+++ myapp/scheduler/scheduler.py
@@ -1,16 +1,19 @@
 from apscheduler.schedulers.background import BackgroundScheduler
 
-from .notification_updater import change_status_rule, CHANGE_STATUS_RULE_THREAD_INT, \
-    CHANGE_THRESHOLD_RULE_THREAD_INT, change_threshold_rule, price_threshold_rule, recommendation_analyst_rule, \
-    PRICE_THRESHOLD_RULE_THREAD_INT, RECOMMENDATION_ANALYST_RULE_THREAD_INT
+from .notification_msgs_updater import change_status_rule, \
+    change_threshold_rule, price_threshold_rule, recommendation_analyst_rule
 from .stock_api_update import stock_api_update
 from myrails.settings import THREAD_INTERVAL
+
+CHANGE_STATUS_RULE_THREAD_INT = 1  # one minute interval
+CHANGE_THRESHOLD_RULE_THREAD_INT = 1  # one minute interval
+PRICE_THRESHOLD_RULE_THREAD_INT = 1  # one minute interval
+RECOMMENDATION_ANALYST_RULE_THREAD_INT = 1  # one minute interval
 
 
 class scheduler:
     def __init__(self):
         self.scheduler_job = BackgroundScheduler()
-        self.scheduler_job.add_job(stock_api_update)
         self.scheduler_job.add_job(stock_api_update, 'interval',
                                    seconds=THREAD_INTERVAL)
         self.scheduler_job.add_job(change_status_rule, 'interval', minutes=CHANGE_STATUS_RULE_THREAD_INT,
